+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Changelog
   * GET `/1.1/queues/<queue_id>/fallbacks`
   * PUT `/1.1/queues/<queue_id>/fallbacks`
 
+* A new API for including contexts inside context has been added:
+
+  * PUT `/1.1/contexts/<context_id>/contexts`
+
 * The entity of a user cannot be changed anymore. The following resource has been removed:
 
   * PUT `/1.1/users/<user_id>/entities/<entity_id>`